requirements.txt
fastapi
uvicorn
pdfplumber
sqlalchemy
psycopg2-binary
python-multipart
huggingface_hub
sqlalchemy
pdfplumber
fitz
google-genai
sendgrid